src_view/requirements.txt
PyYAML
fastapi
uvicorn[standard]
